refactor(home): replace / with UGC explore content; remove /explore
Issue #72 makes the UGC explore the canonical landing page since
the product is fundamentally about user-to-user problem exchange;
official content moves to /learn. Drop the prior dual-section root
(it is a #73 follow-up to design the curated home) and inline the
filter / toolbar / collection grid that used to live at /explore.
The /explore directory is removed without a 301 (PoC, no external
links to preserve).
