+++ src/app/(protected)/page.tsx
@@ -1,117 +1,69 @@
 import { requireAuth } from "@/lib/auth";
 import { getPublishedCollectionsByNewest } from "@/services/collectionService";
-import { getPublishedCourses } from "@/services/courseService";
-import {
-  getCompletedLessonIdsByUser,
-  getCompletedProblemIdsByUser,
-} from "@/services/progressService";
+import { getCompletedProblemIdsByUser } from "@/services/progressService";
+import { BrowseFilterPanel } from "./_components/BrowseFilterPanel";
+import { BrowseToolbar, type SortOption } from "./_components/BrowseToolbar";
 import { CollectionCard } from "./_components/CollectionCard";
-import { CourseCard } from "./_components/CourseCard";
 
 export const dynamic = "force-dynamic";
 
-export default async function Home() {
+const EXPLORE_SORT_OPTIONS: ReadonlyArray<SortOption> = [
+  { key: "newest", label: "新着" },
+  { key: "popular", label: "人気" },
+  { key: "ac-rate", label: "AC率高" },
+  { key: "unchallenged", label: "未挑戦" },
+];
+
+export default async function HomePage() {
   const session = await requireAuth();
-  const [official, community, completedLessons, completedProblems] = await Promise.all([
-    getPublishedCourses(),
+  const [collections, completed] = await Promise.all([
     getPublishedCollectionsByNewest(),
-    getCompletedLessonIdsByUser(session.userId),
     getCompletedProblemIdsByUser(session.userId),
   ]);
-
-  const completedLessonIds = new Set(completedLessons);
-  const completedProblemIds = new Set(completedProblems);
+  const completedIds = new Set(completed);
 
   return (
     <div className="cm-route-enter mx-auto w-full px-6 pt-8 pb-20" style={{ maxWidth: "1280px" }}>
-      <section>
-        <div className="mb-4">
-          <h1 className="m-0 font-semibold text-[22px] tracking-tight">公式コース</h1>
-          <span className="text-xs" style={{ color: "var(--text-3)" }}>
-            codelearn が提供する TypeScript コース
-          </span>
-        </div>
+      <header className="mb-7">
+        <h1 className="m-0 font-bold text-[26px] tracking-tight">探す — コミュニティの問題集</h1>
+        <p className="mt-1.5 text-[13px]" style={{ color: "var(--text-3)" }}>
+          他のユーザーが作った Collection を探そう
+        </p>
+      </header>
 
-        {official.length === 0 ? (
-          <EmptyState variant="official" />
-        ) : (
-          <CardGrid>
-            {official.map((c, idx) => (
-              <li key={c.id}>
-                <CourseCard course={c} index={idx} completedIds={completedLessonIds} />
-              </li>
-            ))}
-          </CardGrid>
-        )}
-      </section>
-
-      <section className="mt-12">
-        <div className="mb-4">
-          <h2 className="m-0 font-semibold text-[22px] tracking-tight">コミュニティコレクション</h2>
-          <span className="text-xs" style={{ color: "var(--text-3)" }}>
-            ユーザーが投稿した新着の問題集
-          </span>
-        </div>
-
-        {community.length === 0 ? (
-          <EmptyState variant="community" />
-        ) : (
-          <CardGrid>
-            {community.map((c, idx) => (
-              <li key={c.id}>
-                <CollectionCard collection={c} index={idx} completedIds={completedProblemIds} />
-              </li>
-            ))}
-          </CardGrid>
-        )}
-      </section>
-    </div>
-  );
-}
-
-function CardGrid({ children }: { children: React.ReactNode }) {
-  return (
-    <ul
-      className="grid gap-4"
-      style={{ gridTemplateColumns: "repeat(auto-fill, minmax(280px, 1fr))" }}
-    >
-      {children}
-    </ul>
-  );
-}
-
-function EmptyState({ variant }: { variant: "official" | "community" }) {
-  if (variant === "official") {
-    return (
-      <div
-        className="rounded-[14px] px-8 py-12 text-center text-[13px]"
-        style={{
-          background: "var(--bg-1)",
-          border: "1px dashed var(--line-3)",
-          color: "var(--text-3)",
-        }}
-      >
-        公式コースがまだありません。
-        <code
-          className="mx-1 rounded px-2 py-0.5 text-[12px]"
-          style={{ background: "var(--bg-2)", color: "var(--accent-solid)" }}
-        >
-          npm run db:seed
-        </code>
-        を実行してください。
+      <div className="grid grid-cols-1 gap-7 md:grid-cols-[240px_minmax(0,1fr)]">
+        <BrowseFilterPanel ariaLabel="フィルタ (探す / UI プレビュー)" />
+        <main className="min-w-0">
+          <BrowseToolbar
+            total={collections.length}
+            sortOptions={EXPLORE_SORT_OPTIONS}
+            activeSortKey="newest"
+          />
+          {collections.length === 0 ? (
+            <div
+              className="rounded-[14px] px-8 py-14 text-center text-[13px]"
+              style={{
+                background: "var(--bg-1)",
+                border: "1px dashed var(--line-3)",
+                color: "var(--text-3)",
+              }}
+            >
+              公開されているコレクションがまだありません。
+            </div>
+          ) : (
+            <ul
+              className="grid gap-4"
+              style={{ gridTemplateColumns: "repeat(auto-fill, minmax(240px, 1fr))" }}
+            >
+              {collections.map((collection, idx) => (
+                <li key={collection.id}>
+                  <CollectionCard collection={collection} index={idx} completedIds={completedIds} />
+                </li>
+              ))}
+            </ul>
+          )}
+        </main>
       </div>
-    );
-  }
-  return (
-    <div
-      className="rounded-[14px] px-8 py-12 text-center text-[13px]"
-      style={{
-        background: "var(--bg-1)",
-        border: "1px dashed var(--line-3)",
-        color: "var(--text-3)",
-      }}
-    >
-      コミュニティコレクションはまだ投稿されていません。
     </div>
   );
 }
